Change format of tests
As part of the work into running tests concurrently, I came across problems that I eventually found a solution for, and it's that environments share variables at the top level of the test. The suggested changes are what I've done in Aikau to fix the problem (lots of search-and-replace, lots of manual updating!!), and functionally have no effect at all on the tests apart from making them safe for running concurrently on a single machine (e.g. vagrant, local). If you only ever intend to do BrowserStack multiple environments, then the fact that they're all spun up on separate VMs may negate this requirement, but I believe this relatively small change is well worth doing in terms of test robustness.